fix error

+++ lib/models/Product.dart
@@ -3,7 +3,7 @@ class Product {
   final String maSp;
   final String tenSp;
   final String maDM;
-  final double donGia;
+  final int donGia;
   final String img;
   Product({required this.img,required this.maSp, required this.donGia, required this.maDM,required this.tenSp});
 }
